requirements.txt
Appium-Python-Client~=4.2.0